refactor: remove `JsonNodeToYaml` & `DirectToYaml`
* I was using them to test various ideas I had with dumping objects to YAML. Their usecases are now solved by `dumpScalar`, `dumpSequence` and `dumpMapping`, `dumpYamlNode` and `dumpCompactNode`
